Adjust packages and artifact names
