+++ README.md
@@ -20,7 +20,7 @@ Features:
 
 The library is primarily intended for adding coloring while doing minimal changes to existing applications. For example, coloring of status codes during unit testing.
 
-For Jackson `2.x` support, see the `jackson2.x` branch. Note that the `jackson2.x` branch is on another artifact (and Java package), so both versions can be used in parallel.
+For Jackson `2.x` support, see the `jackson2.x` branch.
 
 ## License
 [Apache 2.0]
@@ -41,7 +41,7 @@ then add
 ```xml
 <dependency>
     <groupId>org.entur.jackson</groupId>
-    <artifactId>jackson3-syntax-highlight</artifactId>
+    <artifactId>jackson-tools-syntax-highlight</artifactId>
     <version>${jackson-syntax-highlight.version}</version>
 </dependency>
 ```
@@ -63,7 +63,7 @@ ext {
 add
 
 ```groovy
-api ("org.entur.jackson:jackson3-syntax-highlight:${jacksonSyntaxHighlightVersion}")
+api ("org.entur.jackson:jackson-tools-syntax-highlight:${jacksonSyntaxHighlightVersion}")
 ```
 </details>
 
@@ -83,7 +83,7 @@ JsonMapper mapper = JsonMapper.builderWithJackson2Defaults()
 ObjectWriter objectWriter = mapper.writerWithDefaultPrettyPrinter();
 
 // create stream and create generator
-gWriter writer = new StringWriter();
+Writer writer = new StringWriter();
 JsonGenerator delegate = objectWriter.createGenerator(writer);
 SyntaxHighlightingJsonGenerator jsonGenerator = SyntaxHighlightingJsonGenerator.create(delegate);
 
@@ -100,7 +100,7 @@ jsonGenerator.close();
 // print writer contents.
 ```
 
-In addition, the JSON structure can be tracked via [TokenStreamContextListener](src/main/java/org/entur/jackson3/jsh/TokenStreamContextListener.java), for stateful coloring of subtrees.
+In addition, the JSON structure can be tracked via [TokenStreamContextListener](src/main/java/org/entur/jackson/tools/jsh/TokenStreamContextListener.java), for stateful coloring of subtrees.
 
 ## Highlighting an object
 Write a full object using `writeObject`, i.e.
@@ -115,13 +115,13 @@ jsonGenerator.writePOJO(obj);
 
 # History
 
- - 2.x: Support for Jackson 2.
+ - 2.x: Support for Jackson 3.x.
  - 1.1.0: Forked from [jackson-syntax-highlight](https://github.com/skjolber/jackson-syntax-highlight) due to too few maintainers. Support for Jackson 2.x.
 
 [Apache 2.0]:          	http://www.apache.org/licenses/LICENSE-2.0.html
 [issue-tracker]:       	https://github.com/entur/jackson-syntax-highlight/issues
 [Maven]:                http://maven.apache.org/
-[SyntaxHighlighter]:	src/main/java/org/entur/jackson3/jsh/SyntaxHighlighter.java
+[SyntaxHighlighter]:	src/main/java/org/entur/jackson/tools/jsh/SyntaxHighlighter.java
 [Jackson]:				https://github.com/FasterXML/jackson
 [ANSI]:					https://en.wikipedia.org/wiki/ANSI_escape_code
 [JSON]:					https://no.wikipedia.org/wiki/JSON
