+++ pom.xml
@@ -1,9 +1,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.entur.jackson</groupId>
-	<artifactId>jackson3-syntax-highlight</artifactId>
+	<artifactId>jackson-tools-syntax-highlight</artifactId>
 	<version>2.0.0-SNAPSHOT</version>
-	<name>jackson-syntax-highlight</name>
+	<name>jackson-tools-syntax-highlight</name>
 	<description>ANSI syntax highlighting for Jackson output</description>
 	<url>https://github.com/entur/jackson-syntax-highlight</url>
