+++ src/main/java/org/entur/jackson/tools/jsh/SyntaxHighlightingPrettyPrinter.java
@@ -1,4 +1,4 @@
-package org.entur.jackson3.jsh;
+package org.entur.jackson.tools.jsh;
 
 import tools.jackson.core.JacksonException;
 import tools.jackson.core.JsonGenerator;
