+++ src/test/java/org/entur/jackson/tools/jsh/AbstractHighlighterTest.java
@@ -1,4 +1,4 @@
-package org.entur.jackson3.jsh;
+package org.entur.jackson.tools.jsh;
 
 import java.io.IOException;
 import java.io.StringWriter;
