+++ src/test/java/org/entur/jackson/tools/jsh/AnsiSyntaxHighlightTest.java
@@ -1,8 +1,6 @@
-package org.entur.jackson3.jsh;
+package org.entur.jackson.tools.jsh;
 
 import org.junit.jupiter.api.Test;
-
-import org.entur.jackson3.jsh.AnsiSyntaxHighlight;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
