+++ src/test/java/org/entur/jackson/tools/jsh/ReadmeTest.java
@@ -1,4 +1,4 @@
-package org.entur.jackson3.jsh;
+package org.entur.jackson.tools.jsh;
 
 import org.junit.jupiter.api.Test;
 import tools.jackson.core.JsonGenerator;
