+++ src/test/java/org/entur/jackson/tools/jsh/SubtreeTokenStreamContextListener.java
@@ -1,15 +1,21 @@
-package org.entur.jackson3.jsh;
+package org.entur.jackson.tools.jsh;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
 import tools.jackson.core.TokenStreamContext;
 
-public class SingleLineSyntaxHighlighter implements SyntaxHighlighter {
+/**
+ * 
+ * Resolver which returns two different {@linkplain SyntaxHighlighter}s
+ * based on the context location.
+ */
+
+public class SubtreeTokenStreamContextListener implements TokenStreamContextListener, SyntaxHighlighter {
 
 	private SyntaxHighlighter base = new DefaultSyntaxHighlighter();
 	
-	private SyntaxHighlighter redBackground = DefaultSyntaxHighlighter
+	private SyntaxHighlighter numberField = DefaultSyntaxHighlighter
 				.newBuilder()
 				.withBackground(AnsiSyntaxHighlight.BACKGROUND_RED)
 				.build();
@@ -17,9 +23,35 @@ public class SingleLineSyntaxHighlighter implements SyntaxHighlighter {
 	private SyntaxHighlighter delegate = base;
 	
 	public SyntaxHighlighter field(TokenStreamContext context) {
+		if(context.pathAsPointer().toString().equals("/object")) {
+			return numberField;
+		}
+
 		return base;
 	}
 
+	@Override
+	public void startObject(TokenStreamContext outputContext) {
+		this.delegate = field(outputContext);
+	}
+
+	@Override
+	public void endObject(TokenStreamContext outputContext) {
+		// reset
+		this.delegate = base;
+	}
+
+	@Override
+	public void startArray(TokenStreamContext outputContext) {
+		this.delegate = field(outputContext);
+	}
+
+	@Override
+	public void endArray(TokenStreamContext outputContext) {
+		// reset
+		this.delegate = base;
+	}
+	
 	@Override
 	public String forCurlyBrackets() {
 		return delegate.forCurlyBrackets();
@@ -47,11 +79,6 @@ public class SingleLineSyntaxHighlighter implements SyntaxHighlighter {
 
 	@Override
 	public String forFieldName(String value) {
-		if("booleanValue".equals(value)) {
-			this.delegate = redBackground;
-		} else {
-			this.delegate = base;
-		}
 		return delegate.forFieldName(value);
 	}
